taxmst and hsnmst added

+++ app/src/main/java/in/net/maitri/xb/billReports/BillReportActivity.java
@@ -189,8 +189,8 @@ public class BillReportActivity extends AppCompatActivity {
                     }
                     summaryLayout.setVisibility(View.VISIBLE);
                     int items = 0, qty = 0;
-                    double discount = 0;
-                    double netAm = 0;
+                    float discount = 0;
+                    float netAm = 0;
                     mProgressDialog.dismiss();
                     for (int i = 0; i < mGetBillMaster.size(); i++) {
                         SalesMst bm = mGetBillMaster.get(i);
@@ -233,10 +233,10 @@ public class BillReportActivity extends AppCompatActivity {
                 String cashName = mst.getCashName();
                 int internalBillNo = mst.getInternalBillNo();
                 String billDateTime = mst.getDateTime();
-                double discount = mst.getDiscount();
-                double netAmt = mst.getNetAmt();
-                double subTotal = discount + netAmt;
-                double qty = mst.getQty();
+                float discount = mst.getDiscount();
+                float netAmt = mst.getNetAmt();
+                float subTotal = discount + netAmt;
+                float qty = mst.getQty();
 
                 brd = new BillReportDialog(BillReportActivity.this, dbHandler.getDateCount(mGetFromDate), dbHandler.getDateCount(mGetToDate), billNo, billDateTime, mProgressDialog, df.format(discount), netAmt, df.format(subTotal), qty, internalBillNo, custName, cashName, mMessenger, mPrinter);
                 brd.show();
